Add search input in small screen
