+++ src/components/Navbar/Navbar.module.css
@@ -1,10 +1,13 @@
-.navbarContainer {
+.container {
   background-color: white;
   color: black;
+  padding: 15px 30px;
+}
+
+.navbarContainer {
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 15px 30px;
 }
 
 .navLinks a {
