+++ src/components/Navbar/index.tsx
@@ -1,30 +1,13 @@
-import React, { useState } from "react";
-import { Link, useNavigate } from "react-router";
-import { Box, Button, InputAdornment, TextField } from "@mui/material";
-import SearchIcon from "@mui/icons-material/Search";
+import { Link } from "react-router";
+import { Box, Button } from "@mui/material";
 import styles from "./Navbar.module.css";
 import LogoMixmix from "../LogoMixmix";
 import useScroll from "../../hooks/useScroll";
 import NavDrawer from "../NavDawer";
+import SearchInput from "../SearchInput";
 
 const Navbar = () => {
-  const navigate = useNavigate();
   const { scrolled } = useScroll();
-  const [search, setSearch] = useState<string>("");
-
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setSearch(event.target.value);
-  };
-
-  const handleInputKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === "Enter") {
-      const encodedSearch = encodeURIComponent(search.trim());
-      if (encodedSearch) {
-        navigate(`/search?s=${search}`);
-        setSearch("");
-      }
-    }
-  };
 
   return (
     <Box
@@ -37,68 +20,56 @@ const Navbar = () => {
         borderBottom: scrolled ? "1px solid rgb(221, 221, 221)" : "transparent",
       }}
     >
-      <Box className={styles.navbarContainer}>
-        <NavDrawer />
-        <Box
-          sx={{
-            flexGrow: 1,
-            display: "flex",
-            justifyContent: { xs: "center", md: "flex-start" },
-            alignItems: "center",
-          }}
-        >
-          <LogoMixmix />
-        </Box>
-        <Box
-          component="div"
-          className={styles.navLinks}
-          sx={{
-            display: { xs: "none", md: "flex" },
-            alignItems: { md: "center" },
-            justifyContent: { md: "center" },
-          }}
-        >
-          <TextField
-            id="search"
-            placeholder="Buscar..."
-            variant="outlined"
-            size="small"
-            value={search}
-            onChange={handleInputChange}
-            onKeyDown={handleInputKeyDown}
-            slotProps={{
-              input: {
-                startAdornment: (
-                  <InputAdornment position="start">
-                    <SearchIcon />
-                  </InputAdornment>
-                ),
-              },
-            }}
-          />
-          <Link className={styles.link} to="/categories">
-            Categorias
-          </Link>
-          <Link className={styles.link} to="/favorites">
-            Favoritos
-          </Link>
-          <Button
-            variant="contained"
-            size="small"
-            component={Link}
-            to="/"
+      <Box className={styles.container}>
+        <Box className={styles.navbarContainer}>
+          <NavDrawer />
+          <Box
             sx={{
-              backgroundColor: "black",
-              color: "white",
-              textTransform: "none",
-              fontSize: "1rem",
-              "&:focus": {
-                color: "white",
-              },
+              flexGrow: 1,
+              display: "flex",
+              justifyContent: { xs: "center", md: "flex-start" },
+              alignItems: "center",
             }}
           >
-            Inicio
-          </Button>
+            <LogoMixmix />
+          </Box>
+          <Box
+            component="div"
+            className={styles.navLinks}
+            sx={{
+              display: { xs: "none", md: "flex" },
+              alignItems: { md: "center" },
+              justifyContent: { md: "center" },
+            }}
+          >
+            <SearchInput />
+            <Link className={styles.link} to="/categories">
+              Categorias
+            </Link>
+            <Link className={styles.link} to="/favorites">
+              Favoritos
+            </Link>
+            <Button
+              variant="contained"
+              size="small"
+              component={Link}
+              to="/"
+              sx={{
+                backgroundColor: "black",
+                color: "white",
+                textTransform: "none",
+                fontSize: "1rem",
+                "&:focus": {
+                  color: "white",
+                },
+              }}
+            >
+              Inicio
+            </Button>
+          </Box>
+        </Box>
+        <Box sx={{ display: { xs: "block", md: "none" } }}>
+          <SearchInput fullWidth={true} />
         </Box>
       </Box>
     </Box>
